Перенес дата классы с роутами в core и определил там общую модель данных с qr кода
