+++ core/src/main/java/com/mrinsaf/core/data/model/DocumentDetails.kt
@@ -1,0 +1,13 @@
+package com.mrinsaf.core.data.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DocumentDetails(
+    val senderCode: String,
+    val kksCode: String,
+    val workType: String,
+    val documentType: String,
+    val documentVersion: String,
+    val uploadDate: String,
+)
