+++ core/src/main/java/com/mrinsaf/core/data/model/ScreenDestination.kt
@@ -1,0 +1,16 @@
+package com.mrinsaf.core.data.model
+
+import kotlinx.serialization.Serializable
+
+sealed class ScreenDestination() {
+    @Serializable
+    object Scanner: ScreenDestination()
+
+    @Serializable
+    data class DocumentDetailsDestination(
+        val data: DocumentDetails
+    ): ScreenDestination()
+
+    @Serializable
+    data class DocumentReader(val documentTitle: String): ScreenDestination()
+}
